@@ -1,27 +1,25 @@
-﻿using AspNetCoreLocalization;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
 
-namespace AspNetCoreLocalization.Controllers
+namespace AspNetCoreLocalization.Controllers;
+
+[ServiceFilter(typeof(LanguageActionFilter))]
+[Route("api/{culture}/[controller]")]
+public class AboutWithCultureInRouteController : Controller
 {
-    [ServiceFilter(typeof(LanguageActionFilter))]
-    [Route("api/{culture}/[controller]")]
-    public class AboutWithCultureInRouteController : Controller
+    // http://localhost:5000/api/it-CH/AboutWithCultureInRoute
+    // http://localhost:5000/api/fr-CH/AboutWithCultureInRoute
+    private readonly IStringLocalizer<SharedResource> _localizer;
+
+
+    public AboutWithCultureInRouteController(IStringLocalizer<SharedResource> localizer)
     {
-        // http://localhost:5000/api/it-CH/AboutWithCultureInRoute
-        // http://localhost:5000/api/fr-CH/AboutWithCultureInRoute
-        private readonly IStringLocalizer<SharedResource> _localizer;
+        _localizer = localizer;
+    }
 
-
-        public AboutWithCultureInRouteController(IStringLocalizer<SharedResource> localizer)
-        {
-            _localizer = localizer;
-        }
-
-        [HttpGet]
-        public string Get()
-        {
-            return _localizer["Name"];
-        }
+    [HttpGet]
+    public string Get()
+    {
+        return _localizer["Name"];
     }
 }
