@@ -1,73 +1,52 @@
 ﻿using AspNetCoreLocalization.Model;
 using Microsoft.AspNetCore.Mvc;
 
-namespace AspNetCoreLocalization.Controllers
+namespace AspNetCoreLocalization.Controllers;
+
+[Route("api/[controller]")]
+public class BoxesController : Controller
 {
-    [Route("api/[controller]")]
-    public class BoxesController : Controller
+    [HttpGet("{id}")]
+    public IActionResult Get(int id)
     {
-        [HttpGet("{id}")]
-        public IActionResult Get(int id)
-        {
-            if (id == 0)
-            {
-                return NotFound(id);
-            }
+        if (id == 0) return NotFound(id);
 
-            return Ok(new Box() { Id = id, Height = 10, Length = 10, Width=10 });
-        }
+        return Ok(new Box { Id = id, Height = 10, Length = 10, Width = 10 });
+    }
 
-        /// <summary>
-        /// http://localhost:5000/api/boxes?culture=it-CH
-        /// Content-Type: application/json
-        /// 
-        /// { "Id":7,"Height":10,"Width":10,"Length":1000}
-        /// </summary>
-        /// <param name="box"></param>
-        /// <returns></returns>
-        [HttpPost]
-        public IActionResult Post([FromBody]Box box)
-        {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-            else
-            {           
-                string url = Url.RouteUrl("api/boxes", new { id = 11111 },
-                    Request.Scheme, Request.Host.ToUriComponent());
+    /// <summary>
+    ///     http://localhost:5000/api/boxes?culture=it-CH
+    ///     Content-Type: application/json
+    ///     { "Id":7,"Height":10,"Width":10,"Length":1000}
+    /// </summary>
+    /// <param name="box"></param>
+    /// <returns></returns>
+    [HttpPost]
+    public IActionResult Post([FromBody] Box box)
+    {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
 
-                return Created(url, box);
-            }
-        }
+        var url = Url.RouteUrl("api/boxes", new { id = 11111 },
+            Request.Scheme, Request.Host.ToUriComponent());
 
-        [HttpPut("{id}")]
-        public IActionResult Put(int id, [FromBody]Box box)
-        {
-            if(id == 0)
-            {
-                return NotFound(box);
-            }
+        return Created(url, box);
+    }
 
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-            else
-            {
-                return Ok(box);
-            }
-        }
+    [HttpPut("{id}")]
+    public IActionResult Put(int id, [FromBody] Box box)
+    {
+        if (id == 0) return NotFound(box);
 
-        [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
-        {
-            if (id == 0)
-            {
-                return NotFound(id);
-            }
+        if (!ModelState.IsValid) return BadRequest(ModelState);
 
-            return new NoContentResult();
-        }
+        return Ok(box);
+    }
+
+    [HttpDelete("{id}")]
+    public IActionResult Delete(int id)
+    {
+        if (id == 0) return NotFound(id);
+
+        return new NoContentResult();
     }
 }
