@@ -1,22 +1,22 @@
 ﻿using System.Globalization;
-using Microsoft.Extensions.Logging;
 using Microsoft.AspNetCore.Mvc.Filters;
-    
-namespace AspNetCoreLocalization
+using Microsoft.Extensions.Logging;
+
+namespace AspNetCoreLocalization;
+
+public class LanguageActionFilter : ActionFilterAttribute
 {
-    public class LanguageActionFilter : ActionFilterAttribute
+    private readonly ILogger _logger;
+
+    public LanguageActionFilter(ILoggerFactory loggerFactory)
     {
-        private readonly ILogger _logger;
+        _logger = loggerFactory.CreateLogger("LanguageActionFilter");
+    }
 
-        public LanguageActionFilter(ILoggerFactory loggerFactory)
-        {
-            _logger = loggerFactory.CreateLogger("LanguageActionFilter");
-        }
-
-        public override void OnActionExecuting(ActionExecutingContext context)
-        {         
-            string culture = context.RouteData.Values["culture"].ToString();
-            _logger.LogInformation($"Setting the culture from the URL: {culture}");
+    public override void OnActionExecuting(ActionExecutingContext context)
+    {
+        var culture = context.RouteData.Values["culture"].ToString();
+        _logger.LogInformation($"Setting the culture from the URL: {culture}");
 
 #if NET451
             System.Threading.Thread.CurrentThread.CurrentCulture = new CultureInfo(culture);
@@ -25,10 +25,9 @@ namespace AspNetCoreLocalization
             System.Threading.Thread.CurrentThread.CurrentCulture = new CultureInfo(culture);
             System.Threading.Thread.CurrentThread.CurrentUICulture = new CultureInfo(culture);
 #else
-            CultureInfo.CurrentCulture = new CultureInfo(culture);
-            CultureInfo.CurrentUICulture = new CultureInfo(culture);
+        CultureInfo.CurrentCulture = new CultureInfo(culture);
+        CultureInfo.CurrentUICulture = new CultureInfo(culture);
 #endif
-            base.OnActionExecuting(context);
-        }
+        base.OnActionExecuting(context);
     }
 }
