@@ -1,16 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace Localization.SqlLocalizer.DbStringLocalizer
+namespace Localization.SqlLocalizer.DbStringLocalizer;
+
+public class ExportHistory
 {
-    public class ExportHistory
-    {
-        public long Id { get; set; }
+    public long Id { get; set; }
 
-        public DateTime Exported { get; set; }
+    public DateTime Exported { get; set; }
 
-        public string Reason { get; set; }
-    }
+    public string Reason { get; set; }
 }
