@@ -1,13 +1,12 @@
 ﻿using System;
 
-namespace Localization.SqlLocalizer.DbStringLocalizer
-{
-    public class ImportHistory
-    {
-        public long Id { get; set; }
-        
-        public DateTime Imported { get; set; }
+namespace Localization.SqlLocalizer.DbStringLocalizer;
 
-        public string Information { get; set; }
-    }
+public class ImportHistory
+{
+    public long Id { get; set; }
+
+    public DateTime Imported { get; set; }
+
+    public string Information { get; set; }
 }
