@@ -2,19 +2,19 @@
     ViewData["Title"] = "Home Page";
 }
 
-    https://localhost:44392/api/about
+https://localhost:44392/api/about
 
-<br />
-<br />
+<br/>
+<br/>
 
-    https://localhost:44392/api/about?culture=de-CH
+https://localhost:44392/api/about?culture=de-CH
 
-<br />
-<br />
+<br/>
+<br/>
 
-    https://localhost:44392/api/about?culture=en-US
+https://localhost:44392/api/about?culture=en-US
 
-<br />
-<br />
+<br/>
+<br/>
 
-    https://localhost:44392/api/about?culture=fr-CH
+https://localhost:44392/api/about?culture=fr-CH
