@@ -1,17 +1,17 @@
 ﻿<!DOCTYPE html>
 <html>
 <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta charset="utf-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
     <title>@ViewData["Title"] - Localization</title>
 </head>
 <body>
-   
-    <div class="container body-content">
-        @RenderBody()
 
-    </div>
+<div class="container body-content">
+    @RenderBody()
 
-    @RenderSection("Scripts", required: false)
+</div>
+
+@RenderSection("Scripts", false)
 </body>
 </html>
